room reducerとtestを実装
